Tidy up test files organziation
