+++ tests/e2e/e2e_suite_test.go
@@ -1,0 +1,16 @@
+package main_test
+
+import (
+	"testing"
+
+	_ "github.com/mattn/go-sqlite3"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+// Run e2e tests using the Ginkgo runner.
+func TestE2E(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Marmot E2E Suite")
+}
